renaming index to server
