+++ package.json
@@ -2,10 +2,10 @@
   "name": "hipchathooks",
   "version": "2.0.0",
   "description": "a bunch of endpoints to search various apis for hipchat webhooks",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "test": "NODE_ENV=test istanbul cover --report html node_modules/.bin/_mocha --recursive -- -R spec tests/**/*_spec.js",
-    "test-node": "mocha -R spec ./tests/index.js",
+    "test": "npm run test:node",
+    "test:node": "NODE_ENV=test istanbul cover --report html node_modules/.bin/_mocha --recursive -- -R spec tests/**/*_spec.js",
     "start": "node index",
     "dev": "nodemon index"
   },
